It works!

+++ build.sh
@@ -4,23 +4,16 @@ set -e
 
 dlarch="$1"
 if [ "$1" == "amd64" ]; then
-    dlarch="x64_64"
+    dlarch="x86_64"
 fi
 
 frameworkver="$2"
 
-# CLICK_ARCH=$(dpkg-architecture -qDEB_HOST_ARCH)
-CLICK_ARCH="arm64"
+CLICK_ARCH=$(dpkg-architecture -qDEB_HOST_ARCH)
 CLICK_FRAMEWORK=$frameworkver
 
-pkgver=4.2.923
 srcdir=$ROOT
 pkgdir=$INSTALL_DIR
-pkgfile=Beeper-$pkgver-$dlarch.AppImage
-
-if [ -f "$pkgfile" ]; then
-    rm "$pkgfile"
-fi
 
 mkdir -p $pkgdir
 
@@ -28,13 +21,18 @@ mkdir -p $pkgdir
 export PKG_CONFIG_PATH=$pkgdir/lib/pkgconfig:$pkgdir/share/pkgconfig:$PKG_CONFIG_PATH
 export LD_LIBRARY_PATH=$pkgdir/lib:$LD_LIBRARY_PATH
 
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.5/install.sh | bash
-\. "$HOME/.nvm/nvm.sh"
+# Install asar if not already installed, using nvm to install the newest npm version
+if [ ! -f "$HOME/.nvm/versions/node/v24.16.0/bin/asar" ] ; then
+    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.5/install.sh | bash
+    \. "$HOME/.nvm/nvm.sh"
 
-nvm install 24
-npm install @electron/asar
+    nvm install 24
+    npm install -g @electron/asar
+else
+    \. "$HOME/.nvm/nvm.sh"
+fi
 
-DL_URL="https://api.beeper.com/desktop/download/linux/arm64/stable/com.automattic.beeper.desktop"
+DL_URL="https://api.beeper.com/desktop/download/linux/$dlarch/stable/com.automattic.beeper.desktop"
 
 FILENAME=$(
   curl -v -L --no-progress-meter -r 0-1 "$DL_URL" 2>&1 > /dev/null \
@@ -43,17 +41,30 @@ FILENAME=$(
 )
 
 # Pull Beeper AppImage
-wget $DL_URL -O "$FILENAME"
+if [ ! -f ./"$FILENAME" ] ; then
+    wget $DL_URL -O "$FILENAME"
+fi
+
+if ! [ -d runtime ]; then 
+    if [[ "${ARCH}" == "amd64" ]]; then
+        QEMU_ARCH="x86_64";
+    elif [[ "${ARCH}" == "arm64" ]]; then
+        QEMU_ARCH="aarch64";
+    elif [[ "${ARCH}" == "armhf" ]]; then
+        QEMU_ARCH="arm";
+    fi;
+fi
+
 chmod +x ./"$FILENAME"
-./"$FILENAME" --appimage-extract
+qemu-${QEMU_ARCH}-static ./"$FILENAME" --appimage-extract
 
 # fix apprun script
 sed -Ei \
-  's@^(if \[ -z \"\$APPDIR\" ] ; then)$@APPDIR="/'"$INSTALL_DIR"'/beeper"\n\1@' \
+  's@^(if \[ -z \"\$APPDIR\" ] ; then)$@APPDIR="./"\n\1@' \
   "squashfs-root/AppRun"
 
 # apprun script
-install -Dm755 "$srcdir/squashfs-root/AppRun" "$pkgdir/usr/bin/beeper"
+install -Dm755 "squashfs-root/AppRun" "$pkgdir/beeper"
 
 # The app source is now packed into an asar archive (resources/app.asar) with
 # native modules kept in resources/app.asar.unpacked. To patch the source we
@@ -71,7 +82,7 @@ MAIN_DIR="$EXTRACT_DIR/build/main"
 LINUX_CONFIG_FILE=$(grep -lE 'export\{[a-zA-Z0-9_]+ as registerLinuxConfig\};' "$MAIN_DIR"/*.mjs | head -n1)
 if [ -z "$LINUX_CONFIG_FILE" ]; then
   echo "error: could not find file exporting registerLinuxConfig in $MAIN_DIR" >&2
-  return 1
+  exit 1
 fi
 sed -i 's/export{[a-zA-Z0-9_]* as registerLinuxConfig};/const noopFunc=function(){};export{noopFunc as registerLinuxConfig};/' "$LINUX_CONFIG_FILE"
 
@@ -83,6 +94,9 @@ mv app.asar.new "$ASAR_FILE"
 
 cp -r squashfs-root/* "$pkgdir/"
 rm -f "$pkgdir/beepertexts.desktop"   # remove upstream desktop file
+
+# Fix App Icon
+convert "$pkgdir/resources/app.asar.unpacked/build/app-icons/Beeper Squared.png" -trim +repage "$pkgdir/icon.png"
 
 # fix permissions
 chmod -R u+rwX,go+rX,go-w "$pkgdir"
